src/fast_api_db/database/connection.py
import os

from dotenv import load_dotenv
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker

# Find .env file
load_dotenv("app/.env")
# load server details
DB_USER = os.getenv("POSTGRES_USER")
DB_PASSWORD = os.getenv("POSTGRES_PASSWORD")
DB_NAME = os.getenv("POSTGRES_DB")
DB_HOST = os.getenv("POSTGRES_HOST")
DB_PORT = os.getenv("POSTGRES_PORT")

# Get connection string for sqlalchemy
DB_URl = f"postgresql://{DB_USER}:{DB_PASSWORD}@{DB_HOST}:{DB_PORT}/{DB_NAME}"
print(DB_URl)

# Connection
engine = create_engine(DB_URl, echo=True)
engine.connect()
SessionLocal = sessionmaker(bind=engine)
