pyproject.toml
[tool.poetry]
name = "fast-api-db"
version = "0.1.0"
description = "fastapi database project"
authors = ["anakin-1181"]
license = "/"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
gunicorn = "^23.0.0"
uvicorn = "^0.32.1"
debugpy = "^1.8.11"
sqlalchemy = "^2.0.34"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
